Add HttpMultiMap
