+++ airframe-http/src/main/scala/wvlet/airframe/http/HttpHeader.scala
@@ -14,12 +14,6 @@
 package wvlet.airframe.http
 
 object HttpHeader {
-  val empty = HttpHeader(Map.empty)
-
-  case class HeaderEntry(key: String, value: String) {
-    override def toString: String = s"${key}: ${value}"
-  }
-
   final val Accept                        = "Accept"
   final val AcceptCharset                 = "Accept-Charset"
   final val AcceptEncoding                = "Accept-Encoding"
@@ -85,87 +79,4 @@ object HttpHeader {
   final val WwwAuthenticate               = "Www-Authenticate"
   final val xForwardedFor                 = "X-Forwarded-For"
   final val xForwardedProto               = "X-Forwarded-Proto"
-}
-
-import HttpHeader._
-
-trait HttpHeaderAccess { this: HttpHeader =>
-  def get(key: String): Option[String]
-  def getAll(key: String): Seq[String]
-  def isEmpty: Boolean
-  def toSeq: Seq[HeaderEntry]
-}
-
-/**
-  * An immutable multi-map implementation for storing HTTP headers
-  */
-case class HttpHeader(private val map: Map[String, Any]) extends HttpHeaderAccess {
-  override def toString: String = {
-    toSeq.mkString(", ")
-  }
-
-  def set(key: String, value: String): HttpHeader = {
-    this.copy(
-      map = map + (key -> value)
-    )
-  }
-  def add(key: String, value: String): HttpHeader = {
-    val newMap = map
-      .get(key).map { v =>
-        val newValue = v match {
-          case s: String                   => Seq(s, value)
-          case lst: Seq[String @unchecked] => lst +: value
-        }
-        map + (key -> newValue)
-      }.getOrElse {
-        map + (key -> value)
-      }
-
-    this.copy(map = newMap)
-  }
-  def remove(key: String): HttpHeader = {
-    this.copy(map = map - key)
-  }
-
-  override def get(key: String): Option[String] = {
-    map.get(key).flatMap { v: Any =>
-      v match {
-        case s: String                   => Some(s)
-        case lst: Seq[String @unchecked] => Option(lst.head)
-        case null                        => None
-        case _                           => Some(v.toString)
-      }
-    }
-  }
-
-  override def getAll(key: String): Seq[String] = {
-    map
-      .get(key).map { v: Any =>
-        v match {
-          case s: String                   => Seq(s)
-          case lst: Seq[String @unchecked] => lst
-          case null                        => Seq.empty
-          case v                           => Seq(v.toString)
-        }
-      }.getOrElse(Seq.empty)
-  }
-
-  override def isEmpty: Boolean = map.isEmpty
-
-  override def toSeq: Seq[HeaderEntry] = {
-    val b = Seq.newBuilder[HeaderEntry]
-    for ((k, v) <- map) {
-      v match {
-        case s: String =>
-          b += HeaderEntry(k, s)
-        case lst: Seq[String @unchecked] =>
-          b ++= lst.map { x => HeaderEntry(k, x) }
-        case null =>
-        // do nothing
-        case other =>
-          b += HeaderEntry(k, other.toString)
-      }
-    }
-    b.result()
-  }
 }
